refactor: delete store util.py
